chore(r2): update generated types and methods

+++ src/resources/r2/buckets/cors.ts
@@ -1,6 +1,7 @@
 // File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
 
 import { APIResource } from '../../../resource';
+import { isRequestOptions } from '../../../core';
 import * as Core from '../../../core';
 
 export class CORS extends APIResource {
@@ -20,7 +21,7 @@ export class CORS extends APIResource {
     params: CORSUpdateParams,
     options?: Core.RequestOptions,
   ): Core.APIPromise<CORSUpdateResponse> {
-    const { account_id, jurisdiction, ...body } = params;
+    const { account_id = this._client.accountId, jurisdiction, ...body } = params;
     return (
       this._client.put(`/accounts/${account_id}/r2/buckets/${bucketName}/cors`, {
         body,
@@ -48,10 +49,19 @@ export class CORS extends APIResource {
    */
   delete(
     bucketName: string,
-    params: CORSDeleteParams,
+    params?: CORSDeleteParams,
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<CORSDeleteResponse>;
+  delete(bucketName: string, options?: Core.RequestOptions): Core.APIPromise<CORSDeleteResponse>;
+  delete(
+    bucketName: string,
+    params: CORSDeleteParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.APIPromise<CORSDeleteResponse> {
-    const { account_id, jurisdiction } = params;
+    if (isRequestOptions(params)) {
+      return this.delete(bucketName, {}, params);
+    }
+    const { account_id = this._client.accountId, jurisdiction } = params;
     return (
       this._client.delete(`/accounts/${account_id}/r2/buckets/${bucketName}/cors`, {
         ...options,
@@ -78,10 +88,19 @@ export class CORS extends APIResource {
    */
   get(
     bucketName: string,
-    params: CORSGetParams,
+    params?: CORSGetParams,
+    options?: Core.RequestOptions,
+  ): Core.APIPromise<CORSGetResponse>;
+  get(bucketName: string, options?: Core.RequestOptions): Core.APIPromise<CORSGetResponse>;
+  get(
+    bucketName: string,
+    params: CORSGetParams | Core.RequestOptions = {},
     options?: Core.RequestOptions,
   ): Core.APIPromise<CORSGetResponse> {
-    const { account_id, jurisdiction } = params;
+    if (isRequestOptions(params)) {
+      return this.get(bucketName, {}, params);
+    }
+    const { account_id = this._client.accountId, jurisdiction } = params;
     return (
       this._client.get(`/accounts/${account_id}/r2/buckets/${bucketName}/cors`, {
         ...options,
@@ -164,7 +183,7 @@ export interface CORSUpdateParams {
   /**
    * Path param: Account ID.
    */
-  account_id: string;
+  account_id?: string;
 
   /**
    * Body param
@@ -238,7 +257,7 @@ export interface CORSDeleteParams {
   /**
    * Path param: Account ID.
    */
-  account_id: string;
+  account_id?: string;
 
   /**
    * Header param: Jurisdiction where objects in this bucket are guaranteed to be
@@ -251,7 +270,7 @@ export interface CORSGetParams {
   /**
    * Path param: Account ID.
    */
-  account_id: string;
+  account_id?: string;
 
   /**
    * Header param: Jurisdiction where objects in this bucket are guaranteed to be
